+++ reikna/cluda/__init__.py
@@ -1,5 +1,5 @@
 from reikna.cluda.api_discovery import cuda_id, ocl_id, api_ids, supports_api, \
-    supported_api_ids, get_api, cuda_api, ocl_api
+    supported_api_ids, get_api, cuda_api, ocl_api, any_api
 from reikna.cluda.api_tools import find_devices
 from reikna.cluda.kernel import Module, Snippet